+++ src/main/resources/ci3-stubs/MY_Controller.php
@@ -20,6 +20,8 @@
  * @property CI_Parser $parser
  * @property CI_Trackback $trackback
  * @property CI_Cache $cache
+ * @property CI_Encrypt $encrypt
+ * @property CI_Encryption $encryption
  */
 class MY_Controller extends CI_Controller {
 }